deployment/requirements.txt
Django==3.2.15
mysqlclient==2.2.4
gunicorn==21.2.0
django-ckeditor==6.5.1
Pillow==10.2.0 